:lipstick: Refactooring device and device webhook screen
