+++ web/package.json
@@ -11,6 +11,7 @@
     "@mdi/font": "7.4.47",
     "axios": "^1.7.7",
     "core-js": "^3.37.1",
+    "intl-tel-input": "^24.6.0",
     "roboto-fontface": "*",
     "vue": "^3.4.31",
     "vuetify": "^3.6.11"
